+++ src/Presentation/Nop.Web/Factories/IOrderModelFactory.cs
@@ -1,0 +1,17 @@
+﻿using Nop.Core.Domain.Orders;
+using Nop.Core.Domain.Shipping;
+using Nop.Web.Models.Order;
+
+namespace Nop.Web.Factories
+{
+    public partial interface IOrderModelFactory
+    {
+        CustomerOrderListModel PrepareCustomerOrderListModel();
+
+        OrderDetailsModel PrepareOrderDetailsModel(Order order);
+
+        ShipmentDetailsModel PrepareShipmentDetailsModel(Shipment shipment);
+
+        CustomerRewardPointsModel PrepareCustomerRewardPoints(int? page);
+    }
+}
